Add files via upload
Currently, larger refrigerator which has lower energy cost and have more features at similar prices is considered a bargain to the consumer. So, the customers are willing to pay extra for these additional features. The manufactures can improve the sales by taking into consideration the identified factors of the problem and improve on those factors.